test: add vue unit tests
